Refactor multiverse graph
